removed offset from widget as it was incorrect

+++ src/lib.rs
@@ -4,7 +4,7 @@
 //! # use tinybit::events::{events, Event, KeyCode, KeyEvent};
 //! # use tinybit::{
 //! #     term_size, Camera, DebugOutput, Renderer, ScreenPos, ScreenSize, StdoutTarget, Viewport,
-//! #     WorldPos, WorldSize,
+//! #     WorldPos, WorldSize, Pixel
 //! # };
 //! 
 //! fn main() {
@@ -15,6 +15,7 @@
 //!     let mut viewport = Viewport::new(ScreenPos::new(0, 4), viewport_size);
 //! 
 //!     // Camera
+//!     let (width, height) = (width as isize, height as isize);
 //!     let camera_size = WorldSize::new(width / 2, height / 2); let camera_pos =
 //!     WorldPos::new(width, height);
 //!     let mut camera = Camera::new(camera_pos, camera_size);
@@ -29,7 +30,7 @@
 //!     for event in events(20) {
 //!         match event {
 //!             Event::Tick => {
-//!                 let pixel = (player.0, camera.to_screen(player.1));
+//!                 let pixel = Pixel::new(player.0, camera.to_screen(player.1), None);
 //!                 viewport.draw_pixel(pixel);
 //!                 let _ = renderer.render(&mut viewport);
 //! #               break
@@ -49,6 +50,8 @@
 //! }
 //! ```
 
+use serde::{Serialize, Deserialize};
+
 mod camera;
 mod pixelbuffer;
 mod render;
@@ -59,7 +62,7 @@ pub mod events;
 pub mod widgets;
 
 /// A character at a position, with a colour
-#[derive(Debug, Copy, Clone)]
+#[derive(Debug, Copy, Clone, Serialize, Deserialize, PartialEq)]
 pub struct Pixel {
     glyph: char,
     pos: ScreenPos,
@@ -97,25 +100,27 @@ pub use crossterm::style::Color;
 pub type Vec2D<T> = euclid::default::Vector2D<T>;
 
 /// Constraining units to screen space
+#[derive(Serialize, Deserialize, Debug)]
 pub struct Screen;
 
 /// Constraining units to world space
+#[derive(Serialize, Deserialize, Debug)]
 pub struct World;
 
 /// A position on screen, where 0,0 is the top left corner
 pub type ScreenPos = euclid::Point2D<u16, Screen>;
 
 /// A position in the world
-pub type WorldPos = euclid::Point2D<u16, World>;
+pub type WorldPos = euclid::Point2D<isize, World>;
 
 /// A rect on screen
 pub type ScreenRect = euclid::Rect<u16, Screen>;
 
 /// A rect in the world
-pub type WorldRect = euclid::Rect<u16, World>;
+pub type WorldRect = euclid::Rect<isize, World>;
 
 /// A size on screen
 pub type ScreenSize = euclid::Size2D<u16, Screen>;
 
 /// A size in the world
-pub type WorldSize = euclid::Size2D<u16, World>;
+pub type WorldSize = euclid::Size2D<isize, World>;
